+++ Changelog.md
@@ -165,6 +165,9 @@ using the same arguments used in the command line.
     1. Add Insert tab to model editor
         * [Pull request #1924](https://bitbucket.org/osrf/gazebo/pull-request/1924)
 
+    1. Support inserting nested models from model maker
+        * [Pull request #1982](https://bitbucket.org/osrf/gazebo/pull-request/1982)
+
 ## Gazebo 6.0
 
 ### Gazebo 6.5.0 (2015-10-22)